adjust city size in forests
creates some real rural areas

+++ src/overmap.h
@@ -200,7 +200,7 @@ class overmap
         const point_abs_om &pos() const {
             return loc;
         }
-        int get_urbanity(){
+        int get_urbanity() {
             return urbanity;
         }
 
@@ -447,13 +447,13 @@ class overmap
 
         // City Building
         overmap_special_id pick_random_building_to_place( int town_dist ) const;
-        
+
         // urbanity and forestosity are biome stats that can be used to trigger changes in biome.
         int urbanity = 0;
-        // forest_size_adjust is basically the same as forestosity, but forestosity is cast
-        // to an int and scaled to be comparable to urbanity and other biome stats.
+        // forest_size_adjust is basically the same as forestosity, but forestosity is
+        // scaled to be comparable to urbanity and other biome stats.
         float forest_size_adjust = 0.0f;
-        int forestosity = 0;
+        float forestosity = 0.0f;
         void calculate_urbanity();
         void calculate_forestosity();
